+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 # 2.0.0
 ## Feature
++ **[commons-lang]**
+    + ObjectUtils新增isNotEmpty方法
+    + CharUtils新增isNumber方法
+    + CharUtils新增isLetterLower,isLetterUpper, isLetter等方法
 ## Fixed
 ## Change
 + **[commons-json]**
   + 优化json创建以及多json优选
++ **[commons-lang]**
   + 优化使用jdk7提供的安全随机数生成器ThreadLocalRandom
 # 1.0.2
 ## Feature